libs/langgraph-go/worker-py/pyproject.toml
[project]
name = "worker-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = []

[dependency-groups]
dev = [
    "grpcio>=1.71.0",
    "grpcio-tools>=1.71.0",
    "grpclib>=0.4.8",
    "protobuf>=5.29.4",
]
